Переименовывает тесты
