+++ package.json
@@ -50,7 +50,6 @@
     "history": "4.10.1",
     "leaflet": "1.6.0",
     "moment": "2.27.0",
-    "prop-types": "15.7.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-redux": "7.2.0",
