+++ src/components/add-review/add-review.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import renderer from 'react-test-renderer';
+import * as renderer from 'react-test-renderer';
 import {CommentStatus} from '../../constants';
-import {AddReview} from './add-review.tsx';
+import {AddReview} from './add-review';
 
 it(`Render review form`, () => {
   const tree = renderer
