+++ src/components/app/app.test.tsx
@@ -1,21 +1,23 @@
 import MockAdapter from 'axios-mock-adapter';
 import * as React from 'react';
+import * as renderer from 'react-test-renderer';
 import {Provider} from 'react-redux';
-import renderer from 'react-test-renderer';
 import {applyMiddleware, createStore} from 'redux';
 import {composeWithDevTools} from 'redux-devtools-extension';
 import thunk from 'redux-thunk';
 import {createAPI} from '../../api/api';
-import {AppStatus, AuthStatus, CardType} from '../../constants';
+import {AppStatus, AuthStatus} from '../../constants';
 import {setActiveCity, setAppStatus} from '../../reducer/app/app.reducer';
 import {setCities, setNearbyOffers, setOffers, setReviews} from '../../reducer/data/data.reducer';
 import reducer from '../../reducer/reducer';
 import {setAuthStatus, setUserInfo} from '../../reducer/user/user.reducer';
 import {reviews, serverReviews} from '../../test-mocks/reviews';
-import Map from './map.tsx';
+import App from './app';
 import {cities} from '../../test-mocks/cities';
 import {offers, serverOffers} from '../../test-mocks/offers';
 import {serverUserInfo, userInfo} from '../../test-mocks/user';
+import {Router} from 'react-router-dom';
+import {history} from '../../history';
 
 jest.mock(`../map/mapx`, () => `map`);
 
@@ -38,11 +40,15 @@ apiMock
   .onGet(`/login`)
   .reply(200, serverUserInfo);
 
+apiMock
+  .onGet(`/login`)
+  .reply(200, serverUserInfo);
+
 const store = createStore(
     reducer,
     composeWithDevTools(
-        applyMiddleware(thunk.withExtraArgument(api)),
-    ),
+        applyMiddleware(thunk.withExtraArgument(api))
+    )
 );
 
 store.dispatch(setOffers(offers));
@@ -54,18 +60,14 @@ store.dispatch(setAppStatus(AppStatus.SUCCESS_LOAD));
 store.dispatch(setUserInfo(userInfo));
 store.dispatch(setAuthStatus(AuthStatus.AUTH));
 
-it(`Render Map`, () => {
-  const tree = renderer
-    .create(
+it(`App snapshot`, () => {
+  const tree = renderer.create(
+      <Router history={history}>
         <Provider store={store}>
-          <Map
-            type={CardType.MAIN}
-            offers={offers}
-            coordinates={cities[0].coordinates}
-          />
+          <App />
         </Provider>
-    )
-    .toJSON();
+      </Router>
+  ).toJSON();
 
   expect(tree).toMatchSnapshot();
 });
