+++ src/components/cities-list/cities-list.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import renderer from 'react-test-renderer';
-import {CitiesList} from './cities-list.tsx';
+import * as renderer from 'react-test-renderer';
+import {CitiesList} from './cities-list';
 import {cities} from '../../test-mocks/cities';
