+++ src/components/empty-favorites-screen/empty-favorites-screen.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-const EmptyFavoritesScreen: React.FunctionComponent = () =>
+const EmptyFavoritesScreen: React.FunctionComponent<{}> = () =>
   <main className="page__main page__main--favorites page__main--favorites-empty">
     <div className="page__favorites-container container">
       <section className="favorites favorites--empty">
