+++ src/components/empty-offers-screen/empty-offers-screen.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import renderer from 'react-test-renderer';
-import EmptyOffersScreen from './empty-offers-screen.tsx';
+import * as renderer from 'react-test-renderer';
+import EmptyOffersScreen from './empty-offers-screen';
 
 it(`Render empty offers screen`, () => {
   const tree = renderer
