+++ src/components/empty-offers-screen/empty-offers-screen.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-const EmptyOffersScreen: React.FunctionComponent = () =>
+const EmptyOffersScreen: React.FunctionComponent<{}> = () =>
   <div className="cities__places-container cities__places-container--empty container">
     <section className="cities__no-places">
       <div className="cities__status-wrapper tabs__content">
