+++ src/components/main/main.test.tsx
@@ -1,9 +1,9 @@
 import MockAdapter from 'axios-mock-adapter';
 import * as React from 'react';
-import renderer from 'react-test-renderer';
-import {Provider} from 'react-redux';
+import * as renderer from 'react-test-renderer';
 import {applyMiddleware, createStore} from 'redux';
 import {composeWithDevTools} from 'redux-devtools-extension';
+import {Provider} from 'react-redux';
 import thunk from 'redux-thunk';
 import {createAPI} from '../../api/api';
 import {AppStatus, AuthStatus} from '../../constants';
@@ -12,7 +12,7 @@ import {setCities, setNearbyOffers, setOffers, setReviews} from '../../reducer/d
 import reducer from '../../reducer/reducer';
 import {setAuthStatus, setUserInfo} from '../../reducer/user/user.reducer';
 import {reviews, serverReviews} from '../../test-mocks/reviews';
-import App from './app.tsx';
+import Main from './main';
 import {cities} from '../../test-mocks/cities';
 import {offers, serverOffers} from '../../test-mocks/offers';
 import {serverUserInfo, userInfo} from '../../test-mocks/user';
@@ -40,15 +40,11 @@ apiMock
   .onGet(`/login`)
   .reply(200, serverUserInfo);
 
-apiMock
-  .onGet(`/login`)
-  .reply(200, serverUserInfo);
-
 const store = createStore(
     reducer,
     composeWithDevTools(
-        applyMiddleware(thunk.withExtraArgument(api))
-    )
+        applyMiddleware(thunk.withExtraArgument(api)),
+    ),
 );
 
 store.dispatch(setOffers(offers));
@@ -60,14 +56,19 @@ store.dispatch(setAppStatus(AppStatus.SUCCESS_LOAD));
 store.dispatch(setUserInfo(userInfo));
 store.dispatch(setAuthStatus(AuthStatus.AUTH));
 
-it(`App snapshot`, () => {
-  const tree = renderer.create(
-      <Router history={history}>
-        <Provider store={store}>
-          <App />
-        </Provider>
-      </Router>
-  ).toJSON();
+it(`Render Main`, () => {
+  const tree = renderer
+    .create(
+        <Router history={history}>
+          <Provider store={store}>
+            <Main
+              onCardTitleClick={() => {}}
+              onFavoriteClick={() => {}}
+            />
+          </Provider>
+        </Router>
+    )
+    .toJSON();
 
   expect(tree).toMatchSnapshot();
 });
