+++ src/components/not-found/not-found.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import renderer from 'react-test-renderer';
-import NotFound from './not-found.tsx';
+import * as renderer from 'react-test-renderer';
+import NotFound from './not-found';
 import {Router} from 'react-router-dom';
 import {history} from '../../history';
