+++ src/components/public-route/public-route.e2e.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Enzyme, {mount} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import {PublicRoute} from './public-route.tsx';
+import {PublicRoute} from './public-route';
 import {MemoryRouter} from "react-router";
 import AppRoute from '../../routes';
 import {AuthStatus} from '../../constants';
